pythonRaspiLED/src/requirements.txt
google==3.0.0
grpcio==1.44.0
grpcio-tools==1.44.0
protobuf==3.20.1
python-dotenv==0.20.0
